# docs/swagger.yaml
basePath: /api/v1
definitions:
  model.ActionResp:
    properties:
      code:
        example: 200
        type: integer
      data:
        description: 恒为 null
      msg:
        example: 编辑成功
        type: string
    type: object
...
  model.Resign:
    properties:
      created_at:
        type: string
      group_count:
        description: 原老师绑定业务员数（一业务员一群，后端计算）
        type: integer
      id:
        type: integer
      operate_ip:
        type: string
      operator:
        type: string
      original_teacher_dept:
        type: string
      original_teacher_dept_id:
        type: integer
      original_teacher_id:
        type: integer
      original_teacher_name:
        type: string
      remark:
        type: string
      replace_teacher_dept:
        type: string
      replace_teacher_id:
        type: integer
      replace_teacher_name:
        type: string
      salesman_dept:
        type: string
      salesman_name:
        type: string
      transfer_content:
        items:
          type: string
        type: array
      transfer_time:
        type: string
      updated_at:
        type: string
    type: object
  model.ResignAddReq:
    properties:
      original_teacher_id:
        type: integer
      remark:
        type: string
      replace_teacher_id:
        type: integer
      transfer_content:
        description: 白名单 group，非空（传 friend 400）
        items:
          type: string
        type: array
    type: object
  model.ResignListResp:
    properties:
      code:
        example: 200
        type: integer
      data:
        properties:
          count:
            example: 7
            type: integer
          list:
            items:
              $ref: '#/definitions/model.Resign'
            type: array
        type: object
      msg:
        example: success
        type: string
    type: object
  model.Teacher:
    properties:
      account:
        type: string
      avatar:
        type: string
      bind_sales_count:
        description: 相关子查询带出，不落列
        type: integer
      created_at:
        type: string
      dept_id:
        type: integer
      dept_name:
        type: string
      id:
        type: integer
      name:
        type: string
      nickname:
        type: string
      phone:
        type: string
      qualification:
        type: string
      rating:
        description: 种子 1-5，编辑后 0/1/2
        type: integer
      signature:
        type: string
      status:
        description: 输出 "1"/"0"
        type: string
      title:
        type: string
      update_by:
        type: string
      updated_at:
        type: string
      work_no:
        type: string
    type: object
  model.TeacherBindReq:
    properties:
      id:
        description: 老师表 id（接口键名 id，内部字段语义仍为 teacher_id）
        type: integer
      user_ids:
        description: 追加语义，仅新增绑定
        items:
          type: integer
        type: array
    type: object
  model.TeacherBoundResp:
    properties:
      code:
        example: 200
        type: integer
      data:
        example:
        - 1
        items:
          type: integer
        type: array
      msg:
        example: success
        type: string
    type: object
  model.TeacherDetail:
    properties:
      avatar:
        type: string
      id:
        type: integer
      level:
        description: 0 无 / 3 初级 / 5 高级
        type: integer
      nickname:
        type: string
      sign:
        type: string
      title:
        type: string
    type: object
  model.TeacherDetailResp:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/model.TeacherDetail'
      msg:
        example: success
        type: string
    type: object
  model.TeacherListReq:
    properties:
      account:
        description: 模糊
        type: string
      bind_sales_count:
        description: 精确，未填不过滤；0 是有效过滤值
        type: integer
      dept_id:
        description: 精确，null/""/缺省不过滤
        type: integer
      id:
        description: 精确，null/""/缺省不过滤
        type: integer
      name:
        description: 模糊
        type: string
      nickname:
        description: 模糊
        type: string
      page_index:
        description: 默认 1（service 层兜底）
        type: integer
      page_size:
        description: 默认 10，上限 100
        type: integer
      qualification:
        description: 精确（中文）
        type: string
      status:
        description: -1 全部不过滤 / 1 启用 / 0 停用
        type: integer
      title:
        description: 模糊
        type: string
      update_begin_time:
        description: yyyy-MM-dd，与 EndTime 成对生效
        type: string
      update_by:
        description: 模糊
        type: string
      update_end_time:
        type: string
    type: object
  model.TeacherListResp:
    properties:
      code:
        example: 200
        type: integer
      data:
        properties:
          count:
            example: 42
            type: integer
          list:
            items:
              $ref: '#/definitions/model.Teacher'
            type: array
        type: object
      msg:
        example: success
        type: string
    type: object
  model.TeacherOption:
    properties:
      dept_name:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  model.TeacherOptionsResp:
    properties:
      code:
        example: 200
        type: integer
      data:
        items:
          $ref: '#/definitions/model.TeacherOption'
        type: array
      msg:
        example: success
        type: string
    type: object
  model.TeacherSalesListResp:
    properties:
      code:
        example: 200
        type: integer
      data:
        properties:
          count:
            example: 3
            type: integer
          list:
            items:
              $ref: '#/definitions/model.TeacherSalesRow'
            type: array
          pageIndex:
            example: 1
            type: integer
          pageSize:
            example: 5
            type: integer
        type: object
      msg:
        example: success
        type: string
    type: object
  model.TeacherSalesRow:
    properties:
      bind_time:
        type: string
      dept_name:
        type: string
      id:
        type: integer
      nickname:
        type: string
      username:
        type: string
    type: object
  model.TeacherUpdateReq:
    properties:
      avatar:
        type: string
      id:
        type: integer
      level:
        description: 0 无 / 3 初级 / 5 高级（前端 editForm 下拉取值）
        type: integer
      sign:
        type: string
      title:
        type: string
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
  description: |-
    chatSys（老师管理/绑定业务员/离职转移）+ 诊股记录接口。
    统一响应结构 {code, msg, data}；写操作 msg 为约定中文，查询类为 "success"。
    业务接口需 Bearer token（JWT + Redis 白名单，见 PLAN-auth.md）。
  title: im系统诊股 API
  version: "1.0"
paths:
...
  /dxsf/resign/add:
    post:
      consumes:
      - application/json
      description: 原老师绑定业务员全部转移给接替老师（去重合并），姓名/部门等冗余快照由后端回查，group_count 按原老师绑定数计算；原老师无绑定业务员时
        400
      parameters:
      - description: 新增离职转移请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.ResignAddReq'
      produces:
      - application/json
      responses:
        "200":
          description: msg 固定「转移成功」，data 恒为 null
          schema:
            $ref: '#/definitions/model.ActionResp'
        "400":
          description: 请求体非法 / 原老师与接替老师相同 / transfer_content 白名单校验失败 / remark 超过 200
            字符 / 原老师无绑定业务员
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 老师不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 新增离职转移
      tags:
      - 离职转移
  /dxsf/resign/list:
    get:
      consumes:
      - application/json
      description: 分页查询离职转移记录，零值筛选字段不参与过滤；dept_id 匹配原老师部门
      parameters:
      - description: 原老师部门 ID
        in: query
        name: dept_id
        type: integer
      - description: 原老师 ID（精确）
        in: query
        name: original_teacher_id
        type: integer
      - description: 接替老师 ID（精确）
        in: query
        name: replace_teacher_id
        type: integer
      - description: 业务员姓名（模糊）
        in: query
        name: salesman_name
        type: string
      - description: 转移时间起（yyyy-MM-dd，与 transfer_end_time 成对生效）
        in: query
        name: transfer_begin_time
        type: string
      - description: 转移时间止（yyyy-MM-dd）
        in: query
        name: transfer_end_time
        type: string
      - description: 页码（默认 1）
        in: query
        name: page_index
        type: integer
      - description: 页大小（默认 10，上限 100）
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResignListResp'
        "400":
          description: 数字参数格式错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 离职转移列表
      tags:
      - 离职转移
  /dxsf/teacher/bind/salesman:
    post:
      consumes:
      - application/json
      description: 追加语义：仅新增绑定，已存在的绑定保持不变；重复绑定幂等
      parameters:
      - description: 绑定业务员请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.TeacherBindReq'
      produces:
      - application/json
      responses:
        "200":
          description: msg 固定「绑定成功」，data 恒为 null
          schema:
            $ref: '#/definitions/model.ActionResp'
        "400":
          description: 请求体非法
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 老师不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 绑定业务员
      tags:
      - 绑定业务员
  /dxsf/teacher/bind/salesman/list:
    get:
      description: 分页查询指定老师已绑定的业务员（id/用户名/昵称/部门/绑定时间）；data 回显 pageIndex/pageSize
      parameters:
      - description: 老师 ID
        in: query
        name: id
        required: true
        type: integer
      - description: 页码（默认 1）
        in: query
        name: page_index
        type: integer
      - description: 页大小（默认 5，上限 100）
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TeacherSalesListResp'
        "400":
          description: id 缺失或非法
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 老师绑定业务员列表
      tags:
      - 绑定业务员
  /dxsf/teacher/bind/salesman/users:
    get:
      description: 返回全部已绑定业务员 userId（去重平铺数组），供人员树过滤已绑定人员
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TeacherBoundResp'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 全量已绑定业务员
      tags:
      - 绑定业务员
  /dxsf/teacher/detail:
    get:
      description: 编辑弹窗回显（昵称/头衔/评级/头像/签名；列 rating/signature 映射接口 level/sign）
      parameters:
      - description: 老师 ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TeacherDetailResp'
        "400":
          description: id 缺失或非法
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 老师不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 老师详情
      tags:
      - 老师管理
  /dxsf/teacher/edit:
    post:
      consumes:
      - application/json
      description: 编辑头衔/评级/头像/签名（仅这 4 个字段可改，冗余快照字段一律忽略；level 0 无 / 3 初级 / 5 高级）
      parameters:
      - description: 编辑老师请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.TeacherUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: msg 固定「编辑成功」，data 恒为 null
          schema:
            $ref: '#/definitions/model.ActionResp'
        "400":
          description: 请求体非法 / level 非 0/3/5 / 签名超过 200 字符
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 老师不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 编辑老师
      tags:
      - 老师管理
  /dxsf/teacher/list:
    post:
      consumes:
      - application/json
      description: 分页查询老师（筛选条件走 JSON body），未传/null 字段不参与过滤；姓名/账号/昵称/头衔/操作人模糊匹配，部门/ID/认证/状态精确匹配
      parameters:
      - description: 查询条件（数值字段传 null 表示未填）
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.TeacherListReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TeacherListResp'
        "400":
          description: 请求体非法
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 老师列表
      tags:
      - 老师管理
  /dxsf/teacher/options:
    get:
      description: 全量老师选项（含停用，离职转移弹窗用）
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TeacherOptionsResp'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 老师下拉选项
      tags:
      - 老师管理
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    description: 值格式：Bearer {token}，登录接口签发
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"